backend_py/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
PyJWT==2.9.0
python-telegram-bot==21.0.1
python-dotenv==1.0.0


